Make MinecraftLocator respect non-MOD types
Fix #8344
