+++ fmlloader/src/main/java/net/minecraftforge/fml/loading/moddiscovery/AbstractJarFileLocator.java
@@ -43,10 +43,8 @@ import java.util.jar.Manifest;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public abstract class AbstractJarFileLocator implements IModLocator {
+public abstract class AbstractJarFileLocator extends AbstractModLocator {
     private static final Logger LOGGER = LogManager.getLogger();
-    protected static final String MODS_TOML = "META-INF/mods.toml";
-    protected static final String MANIFEST = "META-INF/MANIFEST.MF";
 
     @Override
     public void scanFile(final IModFile file, final Consumer<Path> pathConsumer) {
@@ -70,69 +68,4 @@ public abstract class AbstractJarFileLocator implements IModLocator {
     }
 
     public abstract Stream<Path> scanCandidates();
-
-    @Override
-    public boolean isValid(final IModFile modFile) {
-        return true;
-    }
-
-    private Optional<IModFile> createMod(Path path) {
-        var mjm = new ModJarMetadata();
-        var sj = SecureJar.from(
-            Manifest::new,
-            jar -> jar.findFile(MODS_TOML).isPresent() ? mjm : JarMetadata.from(jar, path),
-            (root, p) -> true,
-            path
-        );
-
-        IModFile mod = null;
-        var type = sj.getManifest().getMainAttributes().getValue(ModFile.TYPE);
-        if (sj.findFile(MODS_TOML).isPresent()) {
-            LOGGER.debug(LogMarkers.SCAN, "Found {} mod of type {}: {}", MODS_TOML, type, path);
-            mod = new ModFile(sj, this, ModFileParser::modsTomlParser);
-        } else if (type != null) {
-            LOGGER.debug(LogMarkers.SCAN, "Found {} mod of type {}: {}", MANIFEST, type, path);
-            mod = new ModFile(sj, this, this::manifestParser);
-        } else {
-            return Optional.empty();
-        }
-
-        mjm.setModFile(mod);
-        return Optional.ofNullable(mod);
-    }
-
-    private IModFileInfo manifestParser(final IModFile mod) {
-        Function<String, Optional<String>> cfg = name -> Optional.ofNullable(mod.getSecureJar().getManifest().getMainAttributes().getValue(name));
-        var license = cfg.apply("LICENSE").orElse("");
-        var dummy = new IConfigurable() {
-            @Override
-            public <T> Optional<T> getConfigElement(String... key) {
-                return Optional.empty();
-            }
-            @Override
-            public List<? extends IConfigurable> getConfigList(String... key) {
-                return Collections.emptyList();
-            }
-        };
-
-        return new IModFileInfo() {
-            @Override public List<IModInfo> getMods() { return Collections.emptyList(); }
-            @Override public List<LanguageSpec> requiredLanguageLoaders() { return Collections.emptyList(); }
-            @Override public boolean showAsResourcePack() { return false; }
-            @Override public Map<String, Object> getFileProperties() { return Collections.emptyMap(); }
-            @Override public String getLicense() { return license; }
-            @Override public IModFile getFile() { return mod; }
-            @Override public IConfigurable getConfig() { return dummy; }
-
-            // These Should never be called as it's only called from ModJarMetadata.version and we bypass that
-            @Override public String moduleName() { return mod.getSecureJar().name(); }
-            @Override public String versionString() { return null; }
-            @Override public List<String> usesServices() { return null; }
-
-            @Override
-            public String toString() {
-                return "IModFileInfo(" + mod.getFilePath() + ")";
-            }
-        };
-    }
 }
